chore : rename feedback domain components
